add transitions
