+++ projects/kiwano/kiwano.vcxproj.filters
@@ -37,15 +37,15 @@
     <Filter Include="2d\animation">
       <UniqueIdentifier>{9314f30d-5742-48b6-94e5-e3b4284106f6}</UniqueIdentifier>
     </Filter>
+    <Filter Include="2d\transition">
+      <UniqueIdentifier>{f70cecd8-6d5b-405d-8466-d3ca2db9b806}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\src\kiwano\2d\Canvas.h">
       <Filter>2d</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\kiwano\2d\Sprite.h">
-      <Filter>2d</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\kiwano\2d\Transition.h">
       <Filter>2d</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\kiwano\core\Resource.h">
@@ -109,9 +109,6 @@
       <Filter>core</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\kiwano\core\Common.h">
-      <Filter>core</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\kiwano\core\Keys.h">
       <Filter>core</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\kiwano\core\Time.h">
@@ -381,15 +378,30 @@
     <ClInclude Include="..\..\src\kiwano\2d\SpriteFrame.h">
       <Filter>2d</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\kiwano\platform\Keys.h">
+      <Filter>platform</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\kiwano\2d\transition\Transition.h">
+      <Filter>2d\transition</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\kiwano\2d\transition\FadeTransition.h">
+      <Filter>2d\transition</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\kiwano\2d\transition\MoveTransition.h">
+      <Filter>2d\transition</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\kiwano\2d\transition\RotationTransition.h">
+      <Filter>2d\transition</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\kiwano\2d\transition\BoxTransition.h">
+      <Filter>2d\transition</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\src\kiwano\2d\Canvas.cpp">
       <Filter>2d</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\kiwano\2d\Sprite.cpp">
-      <Filter>2d</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\src\kiwano\2d\Transition.cpp">
       <Filter>2d</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\kiwano\core\Resource.cpp">
@@ -622,6 +634,21 @@
     </ClCompile>
     <ClCompile Include="..\..\src\kiwano\2d\SpriteFrame.h.cpp">
       <Filter>2d</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\kiwano\2d\transition\Transition.cpp">
+      <Filter>2d\transition</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\kiwano\2d\transition\FadeTransition.cpp">
+      <Filter>2d\transition</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\kiwano\2d\transition\MoveTransition.cpp">
+      <Filter>2d\transition</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\kiwano\2d\transition\RotationTransition.cpp">
+      <Filter>2d\transition</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\kiwano\2d\transition\BoxTransition.cpp">
+      <Filter>2d\transition</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
