+++ src/kiwano/2d/transition/BoxTransition.cpp
@@ -18,70 +18,41 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#pragma once
-#include <kiwano/platform/Keys.h>
-#include <kiwano/event/Event.h>
+#include <kiwano/2d/transition/BoxTransition.h>
 
 namespace kiwano
 {
-KGE_DECLARE_SMART_PTR(KeyEvent);
-KGE_DECLARE_SMART_PTR(KeyDownEvent);
-KGE_DECLARE_SMART_PTR(KeyUpEvent);
-KGE_DECLARE_SMART_PTR(KeyCharEvent);
 
-/**
- * \addtogroup Events
- * @{
- */
-
-/// \~chinese
-/// @brief 键盘事件
-class KGE_API KeyEvent : public Event
+BoxTransition::BoxTransition(Duration duration)
 {
-public:
-    KeyEvent(const EventType& type);
-};
+    SetDuration(duration);
+}
 
-/// \~chinese
-/// @brief 键盘按下事件
-class KGE_API KeyDownEvent : public KeyEvent
+BoxTransition::BoxTransition() {}
+
+void BoxTransition::Init(Stage* prev, Stage* next)
 {
-public:
-    KeyCode code;  ///< 键值
+    Transition::Init(prev, next);
 
-    KeyDownEvent();
-};
+    in_layer_.SetOpacity(0.f);
+}
 
-/// \~chinese
-/// @brief 键盘抬起事件
-class KGE_API KeyUpEvent : public KeyEvent
+void BoxTransition::Update(Duration dt)
 {
-public:
-    KeyCode code;  ///< 键值
+    Transition::Update(dt);
 
-    KeyUpEvent();
-};
-
-/// \~chinese
-/// @brief 键盘字符事件
-class KGE_API KeyCharEvent : public KeyEvent
-{
-public:
-    char value;  ///< 字符
-
-    KeyCharEvent();
-};
-
-template <>
-struct IsSameEventType<KeyEvent>
-{
-    inline bool operator()(const Event* evt) const
+    if (process_ < .5f)
     {
-        return evt->GetType() == KGE_EVENT(KeyDownEvent) || evt->GetType() == KGE_EVENT(KeyUpEvent)
-               || evt->GetType() == KGE_EVENT(KeyCharEvent);
+        out_layer_.SetClipRect(Rect(window_size_.x * process_, window_size_.y * process_,
+                                    window_size_.x * (1 - process_), window_size_.y * (1 - process_)));
     }
-};
-
-/** @} */
+    else
+    {
+        out_layer_.SetOpacity(0.f);
+        in_layer_.SetOpacity(1.f);
+        in_layer_.SetClipRect(Rect(window_size_.x * (1 - process_), window_size_.y * (1 - process_),
+                                   window_size_.x * process_, window_size_.y * process_));
+    }
+}
 
 }  // namespace kiwano
