+++ src/kiwano/event/MouseEvent.h
@@ -19,7 +19,7 @@
 // THE SOFTWARE.
 
 #pragma once
-#include <kiwano/core/Keys.h>
+#include <kiwano/platform/Keys.h>
 #include <kiwano/event/Event.h>
 #include <kiwano/math/Math.h>
