+++ src/kiwano/platform/Input.h
@@ -22,7 +22,7 @@
 #include <array>
 #include <kiwano/core/Common.h>
 #include <kiwano/base/Module.h>
-#include <kiwano/core/Keys.h>
+#include <kiwano/platform/Keys.h>
 #include <kiwano/event/Event.h>
 #include <kiwano/macros.h>
 #include <kiwano/math/Math.h>
